backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Thien Thu Lau V2 Backend"
requires-python = ">=3.11,<4.0"
dependencies = [
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "alembic (>=1.17.2,<2.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "sqlmodel (>=0.0.31,<0.0.32)",
    "fastapi[standard] (>=0.128.0,<0.129.0)",
    "langchain-google-genai (>=4.2.0,<5.0.0)",
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
